middleware: extract planning_tabs.py from schedule_workbook.py — refs #152
Step 4 of 8 in the Issue #152 decomposition. Extraction-only, no behavior
changes.

- scheduling/planning_tabs.py — 11 offline planning-workbook tab renderers
  (Summary, Schedule-Input, Gym-Allocation, Court-Schedule-Sketch +
  scenario builder, Pod-Resource-Estimate + row builder, and the four
  per-tab _annotate_* helpers)
- Nine builder-dependent renderers take the ScheduleWorkbookBuilder as a
  `builder` first param (they call outward into _compute_court_slots,
  _count_racquet_entries, _make_playoff_ids, etc. which stay in the class);
  the class keeps thin wrappers preserving each method's original
  classmethod/instance binding. The two pure functions use staticmethod()
  aliases like Steps 1-3.
- schedule_workbook.py: 5753 -> 4780 lines (-973); 11 now-orphaned imports
  removed
- All 11 extracted bodies verified AST-identical to originals; the
  church_teams_export.py setattr copy loop still resolves all copied
  renderers; 634 tests pass
